Initial Commit
Added the Module to update Rackspace Nova Agent version to the latest.
The latest version for this commit is 1.3.0.2
The module will not perform an update if the installed version equals
the latest.
When an old version of the agent is detected, it will stop the service
of the agent and the updater. make a backup copy of the existing agent
and installs the new new one before restarting the service.